+++ lib/mongoid/locker/wrapper.rb
@@ -8,10 +8,6 @@ elsif Mongoid::Compatibility::Version.mongoid5?
   require 'mongoid/locker/wrapper5'
 elsif Mongoid::Compatibility::Version.mongoid4?
   require 'mongoid/locker/wrapper4'
-elsif Mongoid::Compatibility::Version.mongoid3?
-  require 'mongoid/locker/wrapper3'
-elsif Mongoid::Compatibility::Version.mongoid2?
-  require 'mongoid/locker/wrapper2'
 else
   raise 'incompatible Mongoid version'
 end